wip, add sources approach
