+++ Functions.Templates/Templates/HttpTrigger-CSharp-Isolated-NetCoreVS/HttpTriggerCSharp.cs
@@ -3,7 +3,6 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.Azure.Functions.Worker;
 using Microsoft.Extensions.Logging;
 
-
 namespace Company.Function
 {
     public class HttpTriggerCSharp
