+++ Functions.Templates/Templates/HttpTrigger-CSharp-Isolated/.template.config/template.json
@@ -46,8 +46,30 @@
             "binding": "HostIdentifier"
         }
     },
+    "Framework": {
+        "type": "bind",
+        "binding": "msbuild:TargetFramework",
+        "defaultValue": "net6.0"
+    },
+    "NetFramework": {
+        "type": "computed",
+        "datatype": "bool",
+        "value": "(Framework == \"net48\")"
+    },
+    "NetCore": {
+        "type": "computed",
+        "datatype": "bool",
+        "value": "(Framework == \"net6.0\" || Framework == \"net7.0\" || Framework == \"net8.0\")"
+    },
     "sources": [
-
+        {
+            "condition": "(NetCore)",
+            "source": "../HttpTriggerCSharp.cs"
+        },
+        {
+            "condition": "(NetFramework)",
+            "source": "../HttpTriggerCSharpNetFx.cs"
+        }
     ],
     "primaryOutputs": [
         {
@@ -64,6 +86,18 @@
             "args": {
                 "referenceType": "package",
                 "reference": "Microsoft.Azure.Functions.Worker.Extensions.Http", "version": "3.1.0",
+                "projectFileExtensions": ".csproj"
+            }
+        },
+        {
+            "condition": "(NetCore)",
+            "Description": "Adding Reference to Microsoft.Azure.Functions.Worker.Extensions.Http.AspNetCore Nuget package",
+            "ActionId": "B17581D1-C5C9-4489-8F0A-004BE667B814",
+            "ContinueOnError": "true",
+            "ManualInstructions": [],
+            "args": {
+                "referenceType": "package",
+                "reference": "Microsoft.Azure.Functions.Worker.Extensions.Http.AspNetCore", "version": "1.1.1",
                 "projectFileExtensions": ".csproj"
             }
         },
